Refactor portfolio determination and deployment
Separated the concerns of portfolio determination and deployment in Survey.java by creating a new Portfolios class. Instead of using the switch statements to show different portfolio panels in Survey.java, it now determines the portfolio and directs to corresponding methods in Portfolios.java to create and display the panels. This improves code modularity and readability.