Update layout and styling for portfolio labels and descriptions
The commit includes the removal of Stocks.java file as it's no longer needed. Changes have been made to the layout and styling of the portfolio labels and descriptions in Portfolios.java to make it more concise and brief. Additionally, the fontsize is reduced from 18 to 14 in the description labels for better readability. Furthermore, stocks information is streamlined and unnecessary stock identification numbers in the portfolio labels are removed.